+++ test/Veggerby.Boards.Tests/Core/Tracing/TraceJsonExporterTests.cs
@@ -24,7 +24,7 @@ public class TraceJsonExporterTests
     {
         // Arrange
         var trace = new EvaluationTrace();
-        trace.Add(new TraceEntry(1, "PhaseEnter", "MainPhase", null, null, null, 123UL, 456UL, 789UL));
+        trace.Add(new TraceEntry(1, "PhaseEnter", "MainPhase", null, null, null, null, null, 123UL, 456UL, 789UL));
 
         // Act
         var json = trace.ToJson();
@@ -39,9 +39,9 @@ public class TraceJsonExporterTests
     {
         // Arrange
         var trace = new EvaluationTrace();
-        trace.Add(new TraceEntry(1, "PhaseEnter", "PhaseA", null, null, null, 1, 2, 3));
-        trace.Add(new TraceEntry(2, "RuleEvaluated", "PhaseA", "SomeRule", null, "Valid", 2, 3, 4));
-        trace.Add(new TraceEntry(3, "RuleApplied", "PhaseA", "SomeRule", "SomeEvent", null, 3, 4, 5));
+        trace.Add(new TraceEntry(1, "PhaseEnter", "PhaseA", null, null, null, null, null, 1, 2, 3));
+        trace.Add(new TraceEntry(2, "RuleEvaluated", "PhaseA", "SomeRule", null, "Valid", "OK", 5, 2, 3, 4));
+        trace.Add(new TraceEntry(3, "RuleApplied", "PhaseA", "SomeRule", "SomeEvent", null, null, 5, 3, 4, 5));
 
         // Act
         var json = trace.ToJson();